Fixed altitude on graph, time that is displayed after training, barometer elevation that continued to rise all time
Elevation gain with barometer was measured each 1ms and because of barometer readings that differ by some small amount even when placed on flat surface and not moving, which caused elevation gain to rise by aproximately 100 meters in 1 second. Training duration is fixed on finished training screen. Fixed error when writing training dates to database.